ambassador/requirements.txt
Django
djangorestframework
mysqlclient
django-mysql
django-cors-headers
PyJWT
Faker
django-redis
stripe